requirements.txt
streamlit>=1.24.1
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
